+++ beacon-chain/monitor/process_attestation_test.go
@@ -53,9 +53,9 @@ func setupService(t *testing.T) *Service {
 	}
 	return &Service{
 		config: &ValidatorMonitorConfig{
-			StateGen:          stategen.New(beaconDB),
-			TrackedValidators: trackedVals,
+			StateGen: stategen.New(beaconDB),
 		},
+		trackedValidators:           trackedVals,
 		latestPerformance:           latestPerformance,
 		aggregatedPerformance:       aggregatedPerformance,
 		trackedSyncCommitteeIndices: trackedSyncCommitteeIndices,
