+++ beacon-chain/monitor/process_exit.go
@@ -9,8 +9,10 @@ import (
 // processExitsFromBlock logs the event of one of our tracked validators' exit was
 // included in a block
 func (s *Service) processExitsFromBlock(blk block.BeaconBlock) {
-	s.monitorLock.RLock()
-	defer s.monitorLock.RUnlock()
+
+	s.RLock()
+	defer s.RUnlock()
+
 	for _, exit := range blk.Body().VoluntaryExits() {
 		idx := exit.Exit.ValidatorIndex
 		if s.trackedIndex(idx) {
@@ -25,8 +27,10 @@ func (s *Service) processExitsFromBlock(blk block.BeaconBlock) {
 // processExit logs the event of one of our tracked validators' exit was processed
 func (s *Service) processExit(exit *ethpb.SignedVoluntaryExit) {
 	idx := exit.Exit.ValidatorIndex
-	s.monitorLock.RLock()
-	defer s.monitorLock.RUnlock()
+
+	s.RLock()
+	defer s.RUnlock()
+
 	if s.trackedIndex(idx) {
 		log.WithFields(logrus.Fields{
 			"ValidatorIndex": idx,
