+++ beacon-chain/monitor/process_sync_committee.go
@@ -11,12 +11,13 @@ import (
 
 // processSyncCommitteeContribution logs the event that one of our tracked
 // validators' aggregated sync contribution has been processed.
-// TODO: We do not log if a sync contribution was included in an aggregate (we
-// log them when they are included in blocks)
+// TODO: We do not log if a sync contribution was included in an aggregate (we log them when they are included in blocks)
 func (s *Service) processSyncCommitteeContribution(contribution *ethpb.SignedContributionAndProof) {
 	idx := contribution.Message.AggregatorIndex
-	s.monitorLock.Lock()
-	defer s.monitorLock.Unlock()
+
+	s.Lock()
+	defer s.Unlock()
+
 	if s.trackedIndex(idx) {
 		aggPerf := s.aggregatedPerformance[idx]
 		aggPerf.totalSyncComitteeAggregations++
@@ -37,8 +38,10 @@ func (s *Service) processSyncAggregate(state state.BeaconState, blk block.Beacon
 		log.WithError(err).Error("Cannot get SyncAggregate")
 		return
 	}
-	s.monitorLock.Lock()
-	defer s.monitorLock.Unlock()
+
+	s.Lock()
+	defer s.Unlock()
+
 	for validatorIdx, committeeIndices := range s.trackedSyncCommitteeIndices {
 		if len(committeeIndices) > 0 {
 			contrib := 0
