+++ beacon-chain/monitor/service.go
@@ -4,10 +4,7 @@ import (
 	"sync"
 
 	types "github.com/prysmaticlabs/eth2-types"
-	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
-	"github.com/prysmaticlabs/prysm/beacon-chain/state"
 	"github.com/prysmaticlabs/prysm/beacon-chain/state/stategen"
-	"github.com/prysmaticlabs/prysm/time/slots"
 )
 
 // ValidatorLatestPerformance keeps track of the latest participation of the validator
@@ -40,8 +37,7 @@ type ValidatorAggregatedPerformance struct {
 // monitor service tracks, as well as the event feed notifier that the
 // monitor needs to subscribe.
 type ValidatorMonitorConfig struct {
-	StateGen          stategen.StateManager
-	TrackedValidators map[types.ValidatorIndex]interface{}
+	StateGen stategen.StateManager
 }
 
 // Service is the main structure that tracks validators and reports logs and
@@ -49,10 +45,11 @@ type ValidatorMonitorConfig struct {
 type Service struct {
 	config *ValidatorMonitorConfig
 
-	// monitorLock Locks access to TrackedValidators, latestPerformance, aggregatedPerformance,
+	// Locks access to TrackedValidators, latestPerformance, aggregatedPerformance,
 	// trackedSyncedCommitteeIndices and lastSyncedEpoch
-	monitorLock sync.RWMutex
+	sync.RWMutex
 
+	trackedValidators           map[types.ValidatorIndex]interface{}
 	latestPerformance           map[types.ValidatorIndex]ValidatorLatestPerformance
 	aggregatedPerformance       map[types.ValidatorIndex]ValidatorAggregatedPerformance
 	trackedSyncCommitteeIndices map[types.ValidatorIndex][]types.CommitteeIndex
@@ -61,28 +58,8 @@ type Service struct {
 
 // TrackedIndex returns if the given validator index corresponds to one of the
 // validators we follow.
-// It assumes the caller holds a Lock on the monitorLock
+// It assumes the caller holds the service lock.
 func (s *Service) trackedIndex(idx types.ValidatorIndex) bool {
-	_, ok := s.config.TrackedValidators[idx]
+	_, ok := s.trackedValidators[idx]
 	return ok
-}
-
-// updateSyncCommitteeTrackedVals updates the sync committee assignments of our
-// tracked validators. It gets called when we sync a block after the Sync Period changes.
-func (s *Service) updateSyncCommitteeTrackedVals(state state.BeaconState) {
-	s.monitorLock.Lock()
-	defer s.monitorLock.Unlock()
-	for idx := range s.config.TrackedValidators {
-		syncIdx, err := helpers.CurrentPeriodSyncSubcommitteeIndices(state, idx)
-		if err != nil {
-			log.WithError(err).WithField("ValidatorIndex", idx).Error(
-				"Sync committee assignments will not be reported")
-			delete(s.trackedSyncCommitteeIndices, idx)
-		} else if len(syncIdx) == 0 {
-			delete(s.trackedSyncCommitteeIndices, idx)
-		} else {
-			s.trackedSyncCommitteeIndices[idx] = syncIdx
-		}
-	}
-	s.lastSyncedEpoch = slots.ToEpoch(state.Slot())
 }
